설정 추가중

+++ src/main/kotlin/com/rubok/diary/user/domain/UserType.kt
@@ -4,11 +4,11 @@ import lombok.Getter
 
 @Getter
 class UserType {
-    private val userId: Int? = null
+    private val userId: Long? = null
     private val role: String? = null
     private val useYn: String? = null
 
-    fun toUser(userId: Int?, role: String?, useYn: String?): SecurityUser {
+    fun toUser(userId: Long?, role: String?, useYn: String?): SecurityUser {
         return SecurityUser(userId!!, role!!, useYn!!)
     }
 }
